Refactor export and import services
