+++ src/import-export/export.service.ts
@@ -1,22 +1,20 @@
-import { Injectable, StreamableFile } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { DataType } from './models/data-type.enum';
 import { SettingsExporter } from '../settings/settings.exporter';
 import { Exporter } from './models/exporter.interface';
-import * as json2csv from 'json2csv';
-import * as fs from 'fs/promises';
-import * as os from 'os';
-import * as path from 'path';
-import * as tar from 'tar';
-import { Readable } from 'stream';
 import { UsersExporter } from '../users/users.exporter';
 import { AttributeTypesExporter } from '../catalog/attribute-types/attribute-types.exporter';
-import { dataTypeDependencies } from './models/data-type-dependencies.data';
 import { GenericError } from '../errors/generic.error';
 import { ProductsExporter } from '../catalog/products/products.exporter';
+import { JsonSerializer } from './json-serializer.service';
+import { ZipSerializer } from './zip-serializer.service';
+import { checkDataTypeDependencies } from './data-type.utils';
 
 @Injectable()
 export class ExportService {
   constructor(
+    private jsonSerializer: JsonSerializer,
+    private zipSerializer: ZipSerializer,
     private settingExporter: SettingsExporter,
     private usersExporter: UsersExporter,
     private attributeTypesExporter: AttributeTypesExporter,
@@ -24,31 +22,17 @@ export class ExportService {
   ) {}
 
   async export(data: DataType[], format: 'json' | 'csv') {
-    this.checkDataTypeDependencies(data);
+    checkDataTypeDependencies(data);
     const toExport: Record<string, any[]> = {};
     for (const key of data) {
       toExport[key] = await this.exportCollection(key);
     }
     if (format === 'json') {
-      return await this.serializeJSON(toExport);
+      return await this.jsonSerializer.serialize(toExport);
     } else if (format === 'csv') {
-      return await this.serializeCSV(toExport);
+      return await this.zipSerializer.serialize(toExport);
     } else {
       throw new GenericError('could not serialize export output');
-    }
-  }
-
-  private checkDataTypeDependencies(data: DataType[]) {
-    for (const type of data) {
-      const dependencies = dataTypeDependencies.find((d) => d[0] === type)?.[1];
-      if (!dependencies) {
-        throw new GenericError(`"${type}" is not recognized data type`);
-      }
-      for (const dependency of dependencies) {
-        if (!data.includes(dependency)) {
-          throw new GenericError(`"${type}" depends on "${dependency}"`);
-        }
-      }
     }
   }
 
@@ -60,33 +44,5 @@ export class ExportService {
       [DataType.Products]: this.productsExporter,
     };
     return await exporters[type].export();
-  }
-
-  private async serializeJSON(
-    data: Record<string, any>,
-  ): Promise<StreamableFile> {
-    const parsed = JSON.stringify(data);
-    return new StreamableFile(Readable.from([parsed]), {
-      type: 'application/json',
-    });
-  }
-
-  private async serializeCSV(
-    data: Record<string, any[]>,
-  ): Promise<StreamableFile> {
-    const fields = Object.keys(data);
-    const files: string[] = [];
-    for (const field of fields) {
-      if (data[field].length > 0) {
-        const parsed = json2csv.parse(data[field]);
-        const filePath = path.join(os.tmpdir(), `${field}.csv`);
-        await fs.writeFile(filePath, parsed);
-        files.push(`${field}.csv`);
-      }
-    }
-    const stream = tar.create({ gzip: true, cwd: os.tmpdir() }, files);
-    return new StreamableFile(stream, {
-      type: 'application/gzip',
-    });
   }
 }
