+++ src/import-export/import.controller.spec.ts
@@ -7,6 +7,8 @@ import { UsersImporter } from '../users/users.importer';
 import { AttributeTypesImporter } from '../catalog/attribute-types/attribute-types.importer';
 import { GenericError } from '../errors/generic.error';
 import { ProductsImporter } from '../catalog/products/products.importer';
+import { JsonSerializer } from './json-serializer.service';
+import { ZipSerializer } from './zip-serializer.service';
 
 describe('ImportController', () => {
   let controller: ImportController;
@@ -16,6 +18,8 @@ describe('ImportController', () => {
       controllers: [ImportController],
       providers: [
         ImportService,
+        JsonSerializer,
+        ZipSerializer,
         {
           provide: SettingsImporter,
           useValue: {
